#958: Further clarify deprecation of TAG_CONTAINER_ARRAY
The commit expands on the deprecation reasoning for
PersistentDataType#TAG_CONTAINER_ARRAY to make it more clear for people
migrating their data as to why the data type is deprecated.

The commit also moves the annotation from obsolete to deprecated. The
special handling of this type in the pdc implementation is more annoying
than anything else and should not be used down the line, making
@Deprecated a more proper annotation for this field.

By: Bjarne Koll <lynxplay101@gmail.com>