#993: Deprecate Command's permission message methods as they no longer work
Ever since Mojang's introduction of Brigadier in 1.13, the client is not
sent any commands to which it does not have permission to execute. If a
client executes a command it is not aware of (and consequently does not
have permission for), the client will send itself instead of forwarding
the failure to the server for its response.

By: 2008Choco <hawkeboyz2@hotmail.com>