Only remove worldgen block entity on changed block
The previous PR #10375 removes block entities on setBlock calls to the
WorldGen region to prevent the existence of a block entity on a block
that does not support it as such.

The logic however assumed that such calls would always follow the order
of:
  1. set the base blockstate
  2. create the block entity
  3. finish world gen

This is however not the case for all of world gen, specifically not for
structure templates being placed under water, as these first set the
block state, then configure its block entity (include the chests
lootable) and finally waterlogs the chest by calling another setBlock.

To prevent this logic from deleting the original block entity, this
commit now only removes the block entity found if the blockstate to be
set differs from the previous one, matching LevelChunk#setBlockState.