verify returned suggestion type
since generic types get erased at runtime, java doesn't actually know what is in the returned. List
convert it to a String array, which only accepts the one type, which throws the ArrayStoreException when failed. additionally, since this occurs in BukkitCommandNode, the plugin/commands class is not in the stacktrace.