Push initial datacomponent obfuscation logic
This general principal is that there is no mutation of the patched map, only swapping of the valid values.

Additional work will need to be done to support hiding enchantments whilst keeping the glint, as we either have to append an enchantment, or need to add a glint override.. which kinda defeats the purpose above. Cant add a vanilla enchantment because those can be removed! Further investigation needed.