Fix cancelling PlayerInteractEvent at (0, 0, 0)
When the PlayerInteractEvent gets canceled, the destroyPos field won't be updated, meaning the block won't get broken because the position doesn't match in the check at ServerPlayerGamemode line 261.
However, the default value for destroyPos is BlockPos.ZERO. This leads to the condition returning to true if the player has not mined any blocks before, even if the interactevent was denied.
The issue also happens if someone is allowed to break a block the first time, but is denied the second time, since the destroyPos will stay the same.
This commit fixes #12196 by making the destroyPos nullable instead of using BlockPos.ZERO as a null value.