Update BlockFadeEvent.java
Old wording of:

"Gets the state of the block that will be fading, melting or disappearing"

Sounds as if it will return the same block as .getBlock()

For example, Ice melts and become water. Current wording of "Gets ... the block that will be ... melting" would mean ice. But the method would return water.