Chunk Related Patches
We will need to look into what to do with the identifiers, as previously tickets held onto that but that is gone now.

But overall I inlined the behavior to be identical to FORCED tickets, as that is what it was trying to replicate before.

I kinda just ignore the identifiers for now, as someone else who is more knowledgeable with this can take over.

I also added a diff to chunkmap very intentionally, for whoever gets that patch needs to make sure that method parameter is transfered. (reducedRange equivalent)