Initial draft for tracking affected block pos
The new InsideBlockEffect correctly applies block based effects to
entities when inside.
This means that the block position of the block the entity was affected
by is no longer available in the actually affecting logic.

The commit tracks block positions for blocks that require it by storing
the additional state in the InsideBlockEffectApplier implementation.

Might be expandable to also track lava hurt.