Do not mark plugin tickets as forced; keep correct ticket types
Plugin tickets being marked as forced will mess up the forceload
command.

CHUNK_LOAD and FUTURE_AWAIT are the only tickets that should be
preserved on shutdown - logic that is waiting on these to finish
loading is not capable of handling a failure to load.
The logic in general is also not capable of handling the removal
of the ticket identifier in 1.21.5. It is possible to fix these
shortcomings but Moonrise will resolve them, so why bother.