Don't early return
Instead of returning if slime list is empty, skip the event call and
progress with the rest of the method (which is a noop rn). If for some
reason slimes can normally be removed with size > 1 and no children (via
e.g. future event calls prior), follow up logic should not be skipped
because the transform event cannot be called.