/net/minecraft/world/entity/decoration/
Note that a method was removed on Armorstand for a bug fix that makes literally zero sense, probably very old.
Also added and fixed the logic for leashes, as any entity can hold leashes but anything can be leashed.

Also very sad hack used to detect if the value was set.

/net/minecraft/world/level/block/entity/trialspawner/

/net/minecraft/server/network