fix sponge absorb resetting block entity data of waterlogged blocks
alternative to https://github.com/PaperMC/Paper/commit/320f25cb04f183c18c16595100897622c88ce073, this commit has the side effect of
not including the waterlogged block into the block list for the later event. Generally this logic is fragile and should be handled by
the BlockStateListPopulator in the future.