Fix #11169
This is a rather aggressive behavior change and is mostly an issue with canPickUpLoot. Normally entities that are able to pickup items have this set to true but this is not the case for raiders. Further changing this behavior will cause behavior to be even more different, so it's better to just undo this change until vanilla properly fixes it.