Fix MC-297817
This issue causes the item frame to move on the client erroneously, so we force the entity to resend its position the next tick.

This utilizes a moonrise patch which has a resync flag that can be triggered for the next tick, so we just utilize this. I stub out the method to ensure that it still compiles without the feature patch.