Specify the class loader when loading services
Plugins might be using threads provided by utilities like the common
fork join pool, which does not have the servers classloader around.
Previously this prevented the service loader from finding server
implementations for API types as the threads context class loader did
not point to the server or one of this child class loaders.

To prevent this, instead specify the class loader of the respective API
type, which itself is loaded by the servers class loader.