Add resetPose to Player class
In certain situations, a plugin might have set a fixed pose and now
wishes to reset the pose to a "natural" state. Sure, we could
just use setPose with the current pose and fixed=false, but
you'd have to wait until the next tick for the pose to update,
and sometimes you need to get information from the new pose *now*.

This commit adds a new method for the Player class that clears
the fixed pose state and immediately determines the correct
pose based on the player's current state. Relevant player data,
such as eye height, is immediately updated as well.