Add random tick batching config option
This is a config option where, instead of running random ticking once every tick, you can run it once every *n* ticks, but when randomly ticked, chunks will be ticked *n* times more. While this does affect vanilla behaviour, the nature of random ticking is that this effect is barely noticeable, but it can have a significant impact on TPS.

With 50 bots randomly spread on a server, and a benchmark running for two minutes, setting randomTickScale to 10 reduced the footprint of optimiseRandomDespawn from 6 ms/t to 3.9 ms/t. On a real server with about 90 players, I have observed a reduction in the footprint from 4.04 ms/t to 0.76 ms/t (both with approximately 28,000 chunks loaded, plus or minus 10%), which is a massive TPS saving.