Deprecate the Conversation API (for removal), Metadata API, and PlayerSpawnLocationEvent (for removal)
The Conversation API has been left to rot for the past few years and is considered dead. It does not support components and has had functionality ripped out over the years. Because of this, it is being deprecated for removal.

The metadata API is extremely misleading and causes memory leaks on every server I have seen use it. It just promotes bad practices and is misleading. The PDC system and just having their own solutions is better for developers.

Lastly deprecate PlayerSpawnLocationEvent for removal as events using this hack will be aggressively removed as we move towards better supporting the config API.