Rewrite ticket counter system
The old method of tracking the counts was to copy a
Long2IntOpenHashMap each time a count was removed or added. This
was under the assumption that tickets with counter types were not
being added or removed frequently.

The new system uses ConcurrentLong2LongChainedHashTable instead
to support concurrently updating ticket counters for different
positions.
We copy the map when retrieving the forced chunks, however this
appears to only be used in the ForceLoad command so this shouldn't
be an issue.

Additionally, the ticket counts were not updated if the ticket
were to expire or if the ticket was UNKNOWN and added during
ChunkHolderManager#removeTicket(). These issues are now resolved.