Add a new event for when an entity's scoreboard tags are changed
This adds a new event, `ScoreboardTagsChangeEvent`, which is called when
an entity gets their scoreboard tags changed, through plugins calling
`Entity#addScoreboardTag`, `/tag`, `/data`, etc.

It is fired before the change happens and cannot be cancelled. It
contains the affected entity, which change it is happening (add, remove
or replace), and which tags are part of the change.