Rework API teleportation to better align with Vanilla (#13181)
Co-authored-by: Lulu13022002 <41980282+Lulu13022002@users.noreply.github.com>


This patch includes extensive changes to teleport logic in the API to better represent vanilla behavior. In particular, teleporting with passengers is now supported by default. This also allows for entities to be teleported across dimensions with passengers, except for players, which is still not supported. Entity state related teleportation flags have been deprecated, as that behavior occurs by default now. Additionally, relative teleportation flags are now supported on entities. 
﻿
Most importantly, we no longer hack onto the respawn logic when teleporting players between dimensions in the API. This is a good step towards working on no longer reusing player entities on dimension swap, which is something we are trying to no longer do as it causes a large amount of desync issues on the client/server.
﻿
Entity teleportation via /tp now aligns with vanilla too, which now maintains velocity when teleporting.
﻿
This is a breaking change.