Position riders before API teleportation
The previous teleportation logic would call Entity#positionRider for the
"teleportation" of passengers in the same dimension.
This prevented plugins that aggressively teleport entities on a per-tick
basis from allowing players to get out of sync with their vehicles as
the vanilla teleport logic simply performs a relative repositioning of
the passengers to the vehicle location.

The introduced diff now repositions the passengers onto their vehicle
before computing their relative offset and repositioning.