Reimplement int based gamerule access
Reimplementation of the <=1.21.10 changes to move game rule reads/writes
to a plain array access instead of a hashmap.

The implementation did not fully remove the backing hashmap, instead
it is kept up-to-date on writes. This decision aims to reduce
the diff needed for this change. The backing map is required in the
codec, merging operations and copies.
This means that writes carry the overhead of updating the map, however
they should be rare enough to not actually matter.