Fix PlayerClientLoadedWorldEvent calls
The change during the update would incorrectly spam the event call
as the timer would reach 0 no matter the value of waitingForRespawn.

========================================================================

The server splits the "loaded" state of a player into two states in
1.21.11:
  1. waitingForRespawn
  2. clientLoadedTimeoutTimer

A player is only considered loaded if they are not waiting for a respawn
and their loaded timer is non-positive.

The player joining the world remains similar to the previous logic as
the waitingForRespawn is 'false' and the timer is reduced every tick
or directly set to 0 if the client sends the appropriate packet.

During a respawn however, the states differ from the 1.21.10
implementation. The waitingForRespawn boolean is set to 'true' during
the player death, similar to .10 (which set its clientLoaded
boolean to false).
However, in .10, the timer would immediately start ticking down again
until it reaches 0, which then flipped the boolean.
In .11, the waitingForRespawn is kept as 'true' until the client
requests a respawn via the ServerboundClientCommandPacket.
This flips the waitingForRespawn boolean back to 'false', however bumps
the timer to 60, which prevents the player from being considered loaded.

The logic differs from .10 in that *both* the boolean and the timer are
needed to determine the player loaded state. Previously the timer would
simply toggle the boolean once reaching 0.

The re-implementation of the event call state moves the event call into
ServerGamePacketListenerImpl#markClientLoaded.
To prevent malicious clients from re-triggering the event, the event is
only called if the player is not yet loaded.
As the timer ticking to 0 would however implicitly mark the player as
loaded, the tick down logic instead only ticks down to 1 and then leaves
the final decrease to the markClientLoaded method, which sets the timer
to 0.

This commit notably does not resolve the open issue that the event is
not called for dimension changes.